// src/Module/Users/DTO/Users.DTO.ts
export class UserDTO {
  id?: number;
  imgavatar?: string;
  imageAvartar?: string;
  username?: string;
  email?: string;
  passwords?: string;
  Newpassword?: string;
  RepeatNewPassword?: string;
  phoneNumber?: number;
  birthday?: Date;
  address?: string;
  status?: number;
  role?: number;
  code?: string;
  expirationTime?: number;
  yourcode?: string;
  newpassword?: string;
}